abstract get orders logic
