+++ src/Orders.js
@@ -1,8 +1,7 @@
 import React, { Component } from 'react'
-import { Table } from 'react-bootstrap'
+import { Table  } from 'react-bootstrap'
 import getWeb3 from './utils/getWeb3'
-import { getOrderTracker } from './utils/contracts'
-
+import { getOrders } from './utils/getOrders'
 
 class Orders extends Component {
   constructor(props) {
@@ -11,7 +10,8 @@ class Orders extends Component {
     this.state = {
       web3: null,
       orderTracker: null,
-      orders: []
+      orders: [],
+      showingSales: true
     }
   }
 
@@ -20,46 +20,11 @@ class Orders extends Component {
       this.setState({
         web3: results.web3,
       }, () => {
-        this.getOrderContract()
-      })
-    })
-  }
-
-  getOrderContract() {
-    getOrderTracker(this.state.web3).then(contract => {
-      this.setState({ orderTracker: contract })
-
-      var fetchOrders = contract.NewOrder({buyer: this.state.web3.eth.accounts[0]}, {fromBlock: 0, toBlock: 'latest'})
-      fetchOrders.watch((error, result) => {
-        if (!error) {
-          var orders = this.state.orders
-          const din = result["args"]["DIN"]["c"][0]
-          const quantity = result["args"]["quantity"]["c"][0]
-          const market = result["args"]["market"]
-          const seller = result["args"]["seller"]
-          const orderID = result["args"]["orderID"]["c"][0]
-          const info = result["args"]["info"]
-          const timestamp = parseInt(result["args"]["timestamp"], 10)
-
-          orders.push({din: din, quantity: quantity, market: market, orderID: orderID, info: info, date: this.date(timestamp), seller: seller})
+        getOrders(this.state.web3, {buyer: this.state.web3.eth.accounts[0]}, {fromBlock: 0, toBlock: 'latest'}).then((orders) => {
           this.setState({orders: orders})
-        } else {
-          console.log(error)
-          fetchOrders.stopWatching()
-        }
+        })
       })
     })
-  }
-
-  date(timestamp) {
-    var date = new Date(timestamp*1000)
-
-    var month = date.getUTCMonth() + 1 //months from 1-12
-    var day = date.getUTCDate()
-    var year = date.getUTCFullYear()
-
-    var formattedDate = month + "/" + day + "/" + year
-    return formattedDate
   }
 
 	render() {
@@ -68,6 +33,7 @@ class Orders extends Component {
         <div className="container-orders-table">
           <div className="container-orders-header">
             <h1>Orders</h1>
+
           </div>
 
           <div className="orders-table">
